refactor: move SAML Configuration to bh-shared-ui
