+++ packages/javascript/bh-shared-ui/src/views/index.ts
@@ -18,6 +18,8 @@ export { default as ApiExplorer } from './ApiExplorer';
 
 export { default as UserProfile } from './UserProfile';
 
+export { default as MakeSSOConfiguration } from './SSOConfiguration';
+
 export * from './Explore';
 
 export * from './DataQuality';
